+++ tests/test_cluster.py
@@ -1,4 +1,4 @@
-from iowmongotools import cluster
+from iowmongotools import cluster, app
 
 sample_cluster_config = {
     'mongos': ['mongo-gce-or-1.project.iponweb.net:27017', 'mongo-gce-or-2.project.iponweb.net:27017',
@@ -45,3 +45,50 @@ def test_create_objects_with_partial_config():
     assert 1 == cluster.create_objects(('gce-sc', 'aws-jp'), {'aws-jp': sample_cluster_config})
     assert [v._name for k, v in cluster.Cluster.objects.items()] == ['aws-jp']
     assert [v._declared_config for k, v in cluster.Cluster.objects.items()] == [sample_cluster_config]
+
+
+def test_generate_commands():
+    sample_cluster = cluster.Cluster('cluster_name', sample_cluster_config)
+    actual_description, actual_kwargs = list(), list()
+    for name, cmds in sample_cluster.generate_commands().items():
+        for cmd in cmds:
+            actual_description.append(str(cmd))
+            actual_kwargs.append(cmd.kwargs)
+    description_of_all_commands = [
+        'adding shard mongo-gce-or-1.project.iponweb.net:27019 to cluster_name',
+        'adding shard mongo-gce-or-2.project.iponweb.net:27019 to cluster_name',
+        'adding shard mongo-gce-or-3.project.iponweb.net:27019 to cluster_name',
+        'enabling sharding for database project',
+        'enabling sharding for collection project.cookies by {\'key\': {\'_id\': \'hashed\'}, \'unique\': False}',
+        'enabling sharding for collection project.uuidh by {\'key\': {\'_id\': \'hashed\'}, \'unique\': False}']
+    kwargs_of_all_commands = [
+        {'addshard': 'mongo-gce-or-1.project.iponweb.net:27019', 'name': 'mongo-gce-or-1'},
+        {'addshard': 'mongo-gce-or-2.project.iponweb.net:27019', 'name': 'mongo-gce-or-2'},
+        {'addshard': 'mongo-gce-or-3.project.iponweb.net:27019', 'name': 'mongo-gce-or-3'},
+        {'enableSharding': 'project'},
+        {'shardCollection': 'project.cookies', 'key': {'_id': 'hashed'}, 'unique': False},
+        {'shardCollection': 'project.uuidh', 'key': {'_id': 'hashed'}, 'unique': False}]
+    assert actual_description == description_of_all_commands
+    assert actual_kwargs == kwargs_of_all_commands
+
+
+# requires real pymongo or extending mongomock
+# def test_commands():
+#     config = {'mongos': ['localhost:27017'], 'shards': sample_cluster_config['shards'],
+#               'databases': sample_cluster_config['databases'], 'collections': sample_cluster_config['collections']}
+#     import sys
+#     sys.modules['pymongo'] = __import__('pymongo')
+#     import pymongo
+#     real_cluster = cluster.Cluster('local', config)
+#     real_cluster._api = pymongo.MongoClient(['mongodb://%s' % mongos for mongos in config['mongos']], connect=False)
+#     sys.modules['pymongo'] = __import__('mongomock')
+#     invoker = app.Invoker()
+#     for name, cmds in real_cluster.generate_commands().items():
+#         invoker.add(cmds)
+#     # clean config database
+#     for key in sample_cluster_config:
+#         if key != 'mongos':
+#             real_cluster._api.config[key].drop()
+#     assert real_cluster.actual_config == {'collections': {}, 'databases': {}, 'mongos': [], 'shards': []}
+#     invoker.execute()
+#     assert real_cluster.actual_config == {'collections': {}, 'databases': {}, 'mongos': [], 'shards': []}